rpc: move stdio transport to its own file
